+++ api/mod-record-specifications.yaml
@@ -55,8 +55,3 @@ paths:
     $ref: 'paths/specification-storage/indicator-codes-path.yaml'
   /specification-storage/indicator-codes/{indicatorCodeId}:
     $ref: 'paths/specification-storage/indicator-codes-by-id-path.yaml'
-
-components:
-  schemas:
-    specificationUpdatedEvent:
-      $ref: 'schemas/specificationUpdatedEvent.yaml'
