--- api/schemas/specificationUpdatedEvent.yaml
@@ -1,11 +1,0 @@
-title: Specification Update Event
-type: object
-description: Specification update event body
-properties:
-  specificationId:
-    description: Specification ID
-    format: uuid
-    type: string
-  tenantId:
-    description: Tenant id
-    type: string
