+++ mod-record-specifications-dto/src/main/java/org/folio/rspec/domain/dto/SpecificationUpdatedEvent.java
@@ -1,0 +1,6 @@
+package org.folio.rspec.domain.dto;
+
+import java.util.UUID;
+
+public record SpecificationUpdatedEvent(UUID specificationId, String tenantId) {
+}
