+++ mod-record-specifications-server/src/test/java/org/folio/api/SpecificationStorageApiIT.java
@@ -9,11 +9,9 @@ import static org.folio.support.ApiEndpoints.specificationFieldsPath;
 import static org.folio.support.ApiEndpoints.specificationRulePath;
 import static org.folio.support.ApiEndpoints.specificationRulesPath;
 import static org.folio.support.ApiEndpoints.specificationsPath;
-import static org.folio.support.KafkaUtils.createAndStartTestConsumer;
 import static org.folio.support.TestConstants.BIBLIOGRAPHIC_SPECIFICATION_ID;
 import static org.folio.support.TestConstants.TENANT_ID;
 import static org.folio.support.TestConstants.USER_ID;
-import static org.folio.support.TestConstants.specificationUpdatedTopic;
 import static org.folio.support.builders.FieldBuilder.local;
 import static org.hamcrest.Matchers.everyItem;
 import static org.hamcrest.Matchers.greaterThan;
@@ -26,48 +24,28 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.jayway.jsonpath.JsonPath;
 import java.util.UUID;
-import java.util.concurrent.LinkedBlockingQueue;
 import org.folio.rspec.domain.dto.ErrorCode;
 import org.folio.rspec.domain.dto.Scope;
 import org.folio.rspec.domain.dto.SpecificationRuleDto;
 import org.folio.rspec.domain.dto.SpecificationRuleDtoCollection;
-import org.folio.rspec.domain.dto.SpecificationUpdatedEvent;
 import org.folio.rspec.domain.dto.ToggleSpecificationRuleDto;
 import org.folio.rspec.exception.ResourceNotFoundException;
 import org.folio.spring.testing.extension.DatabaseCleanup;
 import org.folio.spring.testing.type.IntegrationTest;
-import org.folio.support.IntegrationTestBase;
 import org.folio.support.QueryParams;
-import org.junit.jupiter.api.AfterEach;
+import org.folio.support.SpecificationITBase;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 
 @IntegrationTest
 @DatabaseCleanup(tables = FIELD_TABLE_NAME, tenants = TENANT_ID)
-class SpecificationStorageApiIT extends IntegrationTestBase {
+class SpecificationStorageApiIT extends SpecificationITBase {
 
   @BeforeAll
   static void beforeAll() {
     setUpTenant();
-  }
-
-  @BeforeEach
-  void setUp(@Autowired KafkaProperties kafkaProperties) {
-    consumerRecords = new LinkedBlockingQueue<>();
-    container =
-      createAndStartTestConsumer(specificationUpdatedTopic(),
-        consumerRecords, kafkaProperties, SpecificationUpdatedEvent.class);
-  }
-
-  @AfterEach
-  void tearDown() {
-    consumerRecords.clear();
-    container.stop();
   }
 
   @Test
