+++ mod-record-specifications-server/src/test/java/org/folio/api/SpecificationStorageFieldsApiIT.java
@@ -6,11 +6,9 @@ import static org.folio.support.ApiEndpoints.fieldIndicatorsPath;
 import static org.folio.support.ApiEndpoints.fieldPath;
 import static org.folio.support.ApiEndpoints.fieldSubfieldsPath;
 import static org.folio.support.ApiEndpoints.specificationFieldsPath;
-import static org.folio.support.KafkaUtils.createAndStartTestConsumer;
 import static org.folio.support.TestConstants.BIBLIOGRAPHIC_SPECIFICATION_ID;
 import static org.folio.support.TestConstants.TENANT_ID;
 import static org.folio.support.TestConstants.USER_ID;
-import static org.folio.support.TestConstants.specificationUpdatedTopic;
 import static org.folio.support.builders.FieldBuilder.local;
 import static org.folio.support.builders.FieldBuilder.standard;
 import static org.hamcrest.Matchers.everyItem;
@@ -28,28 +26,23 @@ import java.io.UnsupportedEncodingException;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
-import java.util.concurrent.LinkedBlockingQueue;
 import org.folio.rspec.domain.dto.ErrorCode;
-import org.folio.rspec.domain.dto.SpecificationUpdatedEvent;
 import org.folio.rspec.domain.repository.FieldRepository;
 import org.folio.rspec.exception.ResourceNotFoundException;
 import org.folio.spring.FolioModuleMetadata;
 import org.folio.spring.testing.extension.DatabaseCleanup;
 import org.folio.spring.testing.type.IntegrationTest;
-import org.folio.support.IntegrationTestBase;
+import org.folio.support.SpecificationITBase;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.test.web.servlet.MvcResult;
 
 @IntegrationTest
 @DatabaseCleanup(tables = {INDICATOR_TABLE_NAME, FIELD_TABLE_NAME}, tenants = TENANT_ID)
-class SpecificationStorageFieldsApiIT extends IntegrationTestBase {
+class SpecificationStorageFieldsApiIT extends SpecificationITBase {
 
   @Autowired
   private FieldRepository fieldRepository;
@@ -59,20 +52,6 @@ class SpecificationStorageFieldsApiIT extends IntegrationTestBase {
   @BeforeAll
   static void beforeAll() {
     setUpTenant();
-  }
-
-  @BeforeEach
-  void setUp(@Autowired KafkaProperties kafkaProperties) {
-    consumerRecords = new LinkedBlockingQueue<>();
-    container =
-      createAndStartTestConsumer(specificationUpdatedTopic(),
-        consumerRecords, kafkaProperties, SpecificationUpdatedEvent.class);
-  }
-
-  @AfterEach
-  void tearDown() {
-    consumerRecords.clear();
-    container.stop();
   }
 
   @Test
