+++ mod-record-specifications-server/src/test/java/org/folio/api/SpecificationStorageIndicatorsApiIT.java
@@ -7,11 +7,9 @@ import static org.folio.support.ApiEndpoints.codePath;
 import static org.folio.support.ApiEndpoints.fieldIndicatorsPath;
 import static org.folio.support.ApiEndpoints.indicatorCodesPath;
 import static org.folio.support.ApiEndpoints.indicatorPath;
-import static org.folio.support.KafkaUtils.createAndStartTestConsumer;
 import static org.folio.support.TestConstants.BIBLIOGRAPHIC_SPECIFICATION_ID;
 import static org.folio.support.TestConstants.TENANT_ID;
 import static org.folio.support.TestConstants.USER_ID;
-import static org.folio.support.TestConstants.specificationUpdatedTopic;
 import static org.folio.support.builders.FieldBuilder.local;
 import static org.folio.support.builders.FieldBuilder.standard;
 import static org.folio.support.builders.IndicatorBuilder.basic;
@@ -32,29 +30,24 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.UUID;
-import java.util.concurrent.LinkedBlockingQueue;
 import org.folio.rspec.domain.dto.ErrorCode;
 import org.folio.rspec.domain.dto.Scope;
-import org.folio.rspec.domain.dto.SpecificationUpdatedEvent;
 import org.folio.rspec.domain.repository.FieldRepository;
 import org.folio.rspec.domain.repository.IndicatorCodeRepository;
 import org.folio.rspec.domain.repository.IndicatorRepository;
 import org.folio.rspec.exception.ResourceNotFoundException;
 import org.folio.spring.testing.extension.DatabaseCleanup;
 import org.folio.spring.testing.type.IntegrationTest;
-import org.folio.support.IntegrationTestBase;
+import org.folio.support.SpecificationITBase;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.dao.DataIntegrityViolationException;
 
 @IntegrationTest
 @DatabaseCleanup(tables = {INDICATOR_CODE_TABLE_NAME, INDICATOR_TABLE_NAME, FIELD_TABLE_NAME}, tenants = TENANT_ID)
-class SpecificationStorageIndicatorsApiIT extends IntegrationTestBase {
+class SpecificationStorageIndicatorsApiIT extends SpecificationITBase {
 
   @Autowired
   private FieldRepository fieldRepository;
@@ -66,20 +59,6 @@ class SpecificationStorageIndicatorsApiIT extends IntegrationTestBase {
   @BeforeAll
   static void beforeAll() {
     setUpTenant();
-  }
-
-  @BeforeEach
-  void setUp(@Autowired KafkaProperties kafkaProperties) {
-    consumerRecords = new LinkedBlockingQueue<>();
-    container =
-      createAndStartTestConsumer(specificationUpdatedTopic(),
-        consumerRecords, kafkaProperties, SpecificationUpdatedEvent.class);
-  }
-
-  @AfterEach
-  void tearDown() {
-    consumerRecords.clear();
-    container.stop();
   }
 
   @Test
